requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Flask-Mail
python-dotenv
email_validator